+++ resources/views/components/filters/number_range.blade.php
@@ -1,4 +1,5 @@
 <div
+    wire:key="filter-{{ $filter->getKey() }}"
     x-data="{
         minVal: '',
         maxVal: '',